Convert cite links with an associated node into id links
Currently the `org-roam-ui--send-graphdata` function gets all `id` and
`cite` links from the database. They are all sent over the websocket,
but nothing is done to handle the cite links, they don't appear in the
graph. These are not handled because instead of being an (`id` `id`
`type`) tuple, they are (`id` `ref` `type`).

This change uses a join in the DB to get the `id` of the node that `ref`
is associated with. A map function applied to the list of links converts
any cite link whose `ref` has an associated node into an (`id` `id`
`type`) tuple (changing the `type` from `"cite"` to `"id"`). This new
link is correctly handled and appears in the graph.