Infer the implied links represented in the org-mode hierarchy.
The hierarchical nature of org-mode can imply links between nested nodes
that are not explicitly represented in the org-roam db. This change uses
information in the db to infer these links, and adds them as links
denoted with the new "hierarchy" name.

The node properties we use for this are, olp the string representation
of the elisp list that represents the titles of headlines in the output
that make a path to this node, the title, the file, the level, and the
position.

Parent nodes are found by looking for their titles in the `olp` property
of a node in the same file. We ensure parents occur before the children
(filtering out later nodes with the same name). We also track the level
of parent, allowing us to select the most recent parent (highest level)
and the offset between the parent and the child. This lets us filter out
the previous headlines at the same level before the actual subtree this
node is in.

This is a large, complex sql query, but it does me that we do not need
to reparse any files or do any application level post-processing.

I added a `org-roam-ui-hierarchical-links` variable that controls if
these links appear or not.

This query also makes it possible to get links that have all nodes
pointing to their file node. This is done by changing the `MAX(level)`
aggregation to a `MIN(level)`. Is we wanted to make this customizable,
we could remove doing this in the UI level to streamline things.

Note: Currently the UI still still seems to highlight all nodes in a
file when you mouse over the file node, even if you have turned of the
`Link nodes with parent file` option. If this was powered by hierarchy
nodes from the backend, we could probably make fixing this behavior
easier.