fix: display http(s) reflinks too
Consider a node with:

	:ROAM_REFS: http://www.example.net

And another node referencing the page with a (plain) link:

	Lorem ipsum http://www.example.net ...

With the point inside the first node, the roam buffer lists the other
node under the "Reflinks" section as expected, but the roam UI does not
give any clue about the relation between these two.

This commit ensures that the http(s) links that have nodes with
associated refs are converted to id based links of type ref, so that the
UI can handle them.